Add loading state when uploading a file in App.js
This change adds a loading state when uploading a file in App.js. Previously, there was no indication of the upload progress, which could be confusing for the user. Now, when the user selects a file to upload, the loading state is set to "LOADING" and a visual indicator is shown to indicate that the upload is in progress. This improves the user experience by giving them feedback on the progress of the upload.